Se agrega devolucion de codigos http y manejo de errores

+++ src/repositories/cajaRepository.ts
@@ -1,17 +1,35 @@
 import db from '../config/db.js';
+import httpErrors from 'http-errors';
+const { NotFound, BadRequest } = httpErrors;
 
 const cajaRepository = {
   async getAll() {
     const res = await db.query('SELECT * FROM caja ORDER BY id;');
     return res.rows;
   },
+  async getCajabyNombre(nombre: string) {
+    const res = await db.query('SELECT * FROM caja WHERE nombre = $1', [nombre]);
+    if (res.rows.length === 0) {
+      throw new NotFound('Caja no encontrada');
+    }
+    return res.rows[0];
+  },
 
   async getById(id: number) {
     const res = await db.query('SELECT * FROM caja WHERE id = $1', [id]);
+    if (res.rows.length === 0) {
+      throw new NotFound('Caja no encontrada');
+    }
     return res.rows[0];
   },
 
   async create({ nombre, tipo }: { nombre: string; tipo: string }) {
+    // Verificar si el nombre ya existe
+    const existingCaja = await this.getCajabyNombre(nombre);
+    
+    if (existingCaja) {
+      throw new BadRequest('El nombre de la caja ya está en uso');
+    }
     const res = await db.query(
       'INSERT INTO caja (nombre, tipo) VALUES ($1, $2) RETURNING *',
       [nombre, tipo]
@@ -20,6 +38,12 @@ const cajaRepository = {
   },
 
   async update(id: number, { nombre, tipo }: { nombre: string; tipo: string }) {
+    // Verificar si el nombre ya existe
+    const existingCaja = await this.getCajabyNombre(nombre);
+    
+    if (existingCaja) {
+      throw new BadRequest('El nombre de la caja ya está en uso');
+    }
     const res = await db.query(
       'UPDATE caja SET nombre = $1, tipo = $2 WHERE id = $3 RETURNING *',
       [nombre, tipo, id]
@@ -28,6 +52,10 @@ const cajaRepository = {
   },
 
   async remove(id: number) {
+    const res = await db.query('SELECT * FROM caja WHERE id = $1', [id]);
+    if (res.rows.length === 0) {
+      throw new NotFound('Caja no encontrada');
+    }
     await db.query('DELETE FROM caja WHERE id = $1', [id]);
   }
 };
